requirements.txt
Django~=3.1.3
djangorestframework~=3.12.2
clearbit~=0.1.7
pyhunter~=1.7
requests~=2.25.0
PyJWT~=2.0.0a1